zookeeper获取worker id完善，解决偶尔"ConnectionLoss"问题

+++ pay-api-core/src/main/java/com/pay/api/core/service/IIdService.java
@@ -27,5 +27,5 @@ public interface IIdService {
      *
      * @param workerId
      */
-    void setWorkerId(Long workerId);
+    void setWorkerId(Integer workerId);
 }
